requirements.txt
watchdog>=4.0.0
GitPython>=3.1.0
tomli>=2.0.0; python_version < '3.11'
pytest>=7.0.0
pyinstaller>=6.0.0
